moves HelpText and RequirementText to FieldLabel folder
