+++ components/src/index.js
@@ -14,8 +14,8 @@ export { default as NavBar } from "./NavBar/NavBar";
 export { default as NavBarSearch } from "./NavBarSearch/NavBarSearch";
 export { default as HeaderValidation } from "./Validation/HeaderValidation";
 export { default as Field } from "./Field/Field";
-export { default as HelpText } from "./Field/HelpText";
-export { default as RequirementText } from "./Field/RequirementText";
+export { default as HelpText } from "./FieldLabel/HelpText";
+export { default as RequirementText } from "./FieldLabel/RequirementText";
 export { default as Input } from "./Input/Input";
 export { default as Checkbox } from "./Checkbox/Checkbox";
 export { default as Radio } from "./Radio/Radio";
